feat: market api system*
Refactor partially the market system
Adding more info in response
Adding token verification for each request and user status (admin or
not)

+++ src/api/controller/marketController.js
@@ -1,8 +1,18 @@
 const mongoose = require('mongoose');
 const { Market } = require('../../models/marketModels.js');
+const { verifyUserAdmin } = require('../../services/verifyUserStatus.js');
+const idVerification = require('../../services/verifyLengthId.js');
 
 module.exports = {
-    createItem: async (name, img, price, wheightPrice) => {
+    createItem: async (req, res) => {
+
+        const { name, img, price, wheightPrice } = req.body;
+        const token = req.get('token');
+
+        if(!await verifyUserAdmin(token)){
+            return res.status(403).json({type: "error", message: "Sorry you can't do this"});
+        }
+
         const item = new Market({
             _id: mongoose.Types.ObjectId(),
             name: name,
@@ -10,12 +20,8 @@ module.exports = {
             price: price,
             wheightPrice: wheightPrice
         });
-        try{
-            const newItemEntry = await item.save();
-            return newItemEntry;
-        }catch (error){
-            return error;
-        }
+        const newItemEntry = await item.save();
+        return res.status(201).json(newItemEntry);
         
     },
     getAllItems: async () => {
@@ -25,19 +31,31 @@ module.exports = {
         });
         return result;
     },
-    getItem: async (id) => {
+    getItem: async (req, res) => {
+        const { id } = req.params; 
+        idVerified = idVerification(id);
+        if(idVerified){
+            return res.status(409).json(idVerified.message)
+        }
         const result = await Market.findById(id, (err, item) => { 
             if(err){ return err };
             return item; 
         });
-        return result;
+        return res.status(200).json(result);
     },
-    deleteItem: async (id) => {
+    deleteItem: async (req, res) => {
+        const { id } = req.params;
+        const token = req.get('token');
+
+        if(!await verifyUserAdmin(token)){
+            return res.status(403).json({type: "error", message: "Sorry you can't do this"});
+        }
+
         const result = await Market.findByIdAndDelete(id, (err, item) => {
             if(err){ return err};
             return item;
         });
-        return result;
+        return res.status(200).json(result);
     },
     updateItem: async (id, body) => {
         const result = await Market.findByIdAndUpdate(id, body, (err, item) => {
